Fix 3D sequence problems
Fixed (1) The ADC delay and RO gradient ramp mismatch (now using 40 us for both) for all three 3D sequences and (2) The issue of slew rate not being constrained properly (caused by a post-hoc waveform scaling and left unchecked by PyPulseq itself) in the 3D UTE full pulse option. The sequences are yet to be tested on a scanner.